refactor: Change get to content.
I thought using get for this page was simply just bad. It's about expanding multiple collection helpers. Not just one.
