+++ packages/utilities/src/index.ts
@@ -2,5 +2,5 @@
 export { useTemplaterAndProjector } from "./lib/useTemplaterAndProjector.js"
 export * from "./lib/helpers/conditional"
 export * from "./lib/types"
-export * from "./lib/helpers/get-collection"
+export * from "./lib/helpers/content-collection"
 export * from "./lib/helpers"
